+++ Sources/BaseFoundationiOS/Chains/DefaultOKChainResponse.swift
@@ -1,0 +1,11 @@
+//
+//  DefaultOKChainResponse.swift
+//
+//
+//  Created by pablo borquez on 12/12/23.
+//
+
+import Foundation
+
+open class DefaultOKChainResponse: OKChainResponse<[String: Any]> {
+}