+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `isEmail` func implemented on `StringExtension`.
 - `FormViewValidatorDelegate` implemented.
 - `DateExtension` and `DateFormatterExtension` methos implemented.
+- `DefaultOKChainResponse` implemented.
 
 # 0.11.0